Crate libtest

Macros

CPU_ID

Structs

Stats

Constants

THRESHOLD_ERROR_RATIO

Functions

calculate_stats
Helper function to calculate statistics of a provided dataset

check_myrq
True if only two tasks are running in the current runqueue. Used to verify if there are any other tasks than the current task and idle task in the runqueue

cycle_count_overhead
Measures the overhead of using the PMU reference cycles counter. Calls timing_overhead_inner multiple times and averages the value. Overhead is in reference cycles.

hpet_2_ns
Helper function to convert ticks to nano seconds

hpet_2_us
Helper function to convert ticks to micro seconds

hpet_timing_overhead
Measures the overhead of using the hpet timer. Calls timing_overhead_inner multiple times and averages the value. Overhead is a count value. It is not time.

nr_tasks_in_rq
Helper function return the tasks in a given core’s runqueue

pick_free_core
Helper function to pick a free child core if possible

start_counting_reference_cycles
Starts the PMU counter to measure reference cycles. The PMU should be initialized before calling this function. The PMU initialization, start count and stop count should all be called on the same core.

stop_counting_reference_cycles
Stops the PMU counter and stores the reference cycles since the start. The PMU should be initialized before calling this function. The PMU initialization, start count and stop count should all be called on the same core.